requirements.txt
pyyaml>=6.0
requests>=2.31.0
pytest>=7.4.0